+++ packages/viewer/src/services.ts
@@ -1,4 +1,5 @@
 import Parser from '@wiris/mathtype-html-integration-devkit/src/parser';
+import { Properties } from './properties';
 
 enum MethodType {
   Post = "POST",
@@ -49,11 +50,17 @@ export async function processJsonResponse(response: Promise<Response>): Promise<
 export async function callService(query: object, serviceName: string, method: MethodType, serverURL: string, extension: string) : Promise<any> {
   try {
     const url = new URL(serviceName + extension, serverURL);
+
+    const properties = Properties.getInstance();
+    const headers = {
+      'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+      ...properties.config.backendConfig.wiriscustomheaders
+    }
+
+    // const test = Properties.
     const init: RequestInit = {
       method,
-      headers: {
-        'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
-      },
+      headers
     };
 
     if (method === MethodType.Get) {
@@ -113,7 +120,7 @@ export async function showImage(mml: string, lang: string, url: string, extensio
   }
 
   // Try to obtain the image via GET
-  const getParams = Parser.createShowImageSrcData({ mml }, lang);
+  const getParams = Parser.createShowImageSrcData(params, params.lang);
   const getResponse = callService(getParams, 'showimage', MethodType.Get, url, extension);
   try {
     return await processJsonResponse(getResponse);